- PRNG seeds are now UInt64 everywhere, previously Int32. - Use RandomSourceFactory.Create() instead of new XorShiftRandom(). - Added DiscreteDistribution unit test.

+++ Redzen/Random/IRandomSource.cs
@@ -18,86 +18,65 @@ namespace Redzen.Random
     public interface IRandomSource
     {
         /// <summary>
-        /// Re-initialises using an int value as a seed.
+        /// Re-initialises the random number generator state using the provided seed value.
         /// </summary>
-        void Reinitialise(int seed);
+        void Reinitialise(ulong seed);
 
         /// <summary>
-        /// Generates a random int over the range 0 to int.MaxValue-1.
-        /// MaxValue is not generated in order to remain functionally equivalent to System.Random.Next().
-        /// This does slightly eat into some of the performance gain over System.Random, but not much.
-        /// For better performance see:
-        /// 
-        /// Call NextInt() for an int over the range 0 to int.MaxValue.
-        /// 
-        /// Call NextUInt() and cast the result to an int to generate an int over the full Int32 value range
-        /// including negative values. 
+        /// Generates a random Int32 over the interval [0, int.MaxValue), i.e. exclusive of Int32.MaxValue.
         /// </summary>
         int Next();
 
         /// <summary>
-        /// Generates a random int over the range 0 to upperBound-1, and not including upperBound.
+        /// Generates a random Int32 over the interval [range 0 to upperBound), i.e. excluding upperBound.
         /// </summary>
         int Next(int upperBound);
 
         /// <summary>
-        /// Generates a random int over the range lowerBound to upperBound-1, and not including upperBound.
+        /// Generates a random Int32 over the interval [lowerBound, upperBound), i.e. excluding upperBound.
         /// upperBound must be >= lowerBound. lowerBound may be negative.
         /// </summary>
         int Next(int lowerBound, int upperBound);
 
         /// <summary>
-        /// Generates a random double. Values returned are over the range [0, 1). That is, inclusive of 0.0 and exclusive of 1.0.
+        /// Generates a random double over the interval [0, 1), i.e. inclusive of 0.0 and exclusive of 1.0.
         /// </summary>
         double NextDouble();
 
         /// <summary>
         /// Fills the provided byte array with random bytes.
-        /// This method is functionally equivalent to System.Random.NextBytes(). 
         /// </summary>
+        /// <param name="buffer">The byte array to fill with random values.</param>
         void NextBytes(byte[] buffer);
 
         /// <summary>
-        /// Generates a random float. Values returned are over the range [0, 1). That is, inclusive of 0.0 and exclusive of 1.0.
+        /// Generates a random float over the interval [0, 1), i.e. inclusive of 0.0 and exclusive of 1.0.
         /// </summary>
         float NextFloat();
 
         /// <summary>
-        /// Generates a uint. Values returned are over the full range of a uint, 
-        /// uint.MinValue to uint.MaxValue, inclusive.
-        /// 
-        /// This is the fastest method for generating a single random number because the underlying
-        /// random number generator algorithm generates 32 random bits that can be cast directly to 
-        /// a uint.
+        /// Generates a random UInt32 over the interval [uint.MinValue, uint.MaxValue], i.e. over the full 
+        /// range of a UInt32.
         /// </summary>
         uint NextUInt();
 
         /// <summary>
-        /// Generates a random int over the range 0 to int.MaxValue, inclusive. 
-        /// This method differs from Next() only in that the range is 0 to int.MaxValue
-        /// and not 0 to int.MaxValue-1.
-        /// 
-        /// The slight difference in range means this method is slightly faster than Next()
-        /// but is not functionally equivalent to System.Random.Next().
+        /// Generates a random Int32 over interval [0 to Int32.MaxValue], i.e. inclusive of Int32.MaxValue.
         /// </summary>
         int NextInt();
 
         /// <summary>
-        /// Generates a random double. Values returned are over the range (0, 1). That is, exclusive of both 0.0 and 1.0.
+        /// Generates a random double over the interval (0, 1), i.e. exclusive of both 0.0 and 1.0
         /// </summary>
         double NextDoubleNonZero();
 
         /// <summary>
         /// Generates a single random bit.
-        /// This method's performance is improved by generating 32 bits in one operation and storing them
-        /// ready for future calls.
         /// </summary>
         bool NextBool();
 
         /// <summary>
-        /// Generates a single random byte with range [0,255].
-        /// This method's performance is improved by generating 4 bytes in one operation and storing them
-        /// ready for future calls.
+        /// Generates a single random byte over the interval [0,255].
         /// </summary>
         byte NextByte();
     }
